backend/app/requirements.txt
asyncpg
fastapi
fastapi-cors
uvicorn
sqlalchemy
databases
pydantic
python-dotenv
pytest-html 
coverage 
pytest-cov
requests
psutil
asyncpg
pydantic[email]
passlib
pyjwt
greenlet
boto3
python-multipart